Move files to more structured folder
